+++ packages/create-mastra/CHANGELOG.md
@@ -1,5 +1,9 @@
 # create-mastra
 
+## 1.3.16-alpha.1
+
+## 1.3.16-alpha.0
+
 ## 1.3.15
 
 ### Patch Changes